48234371 Fix up implementation of PUT, add test for successful put, fix column definition bug for Sqlite.
The column definition issue is due to the fact that sqlite doesn't have
a server. So, defining server default values for columns doesn't work.
Rather than having hacky solutions in the service, there is a bit of a
hack in the models to decide on the style of defaulting to use dependent
upon whether or not the database is sqlite.