49871451 Add query filtering to resource collections.
Query filtering by attribute is now supported for all basic attributes
for all ggrc resource collections. However, 'in' or
'contains' type queries aren't supported at this time for attributes
that are lists. It is possible to query for direct references by using
the 'thing_id' parameter, though. For example, to get the Controls with
a specific Directive, supply the query parameter 'directive_id=123'.

Query parameters can be anded together with the '&' operator. For
example, '/api/controls?active=True&directive=123'